requirements.txt
openai
transformers
torch 
pandas
numpy
langchain
streamlit 
tiktoken 
duckduckgo-search
python-dotenv
langchain